Functional block and kill rewards
Made it so that block now actually uses the player's energy and added
kill rewards for killing enemies(shocking I know), pretty much, each
enemy will have an energy reward and a health reward. The energy reward
is given until you are full on energy, at which point you gain the
health reward as many times as spare energy there was in the energy
reward. example: if you have 2/3 energy and get a reward with 3 energy
and 10 health, then you would gain energy and go to 3/3, then the
remaining two energy rewards would be multiplied and given as the health
reward, so in this case, you would get 20 health because 10*2