More console stuffs
Added a bunch more commands, also made it so that the console now tells
you when the connection is made rather than literally nothing. Also,
also, made it so that the system for sending and recieving text from the
console is now better and works consistently. If you want to see the
commands I added, look in the files, you're a smart cookie and I can't
be assed to put them here. They are mostly self explanatory, except for
one on the condition that you play TF2, in which case that singular one
is confusing.