Fixed(?) some visual things
I made it so that instead of one big clear then drawing everthing, it
now clears and draws line by line, making it so that there is less of
v-sync type artifacts. Also made it so that the player now gets checked
that they are on the screen before being rendered.