Fixed bug with opening door with movement keys
It didn't reflood before so enemies just wouldn't see you if you opened
a door like that. Also made it so that you can run commands from a file
the moment the console opens. As a bit of an extra, there is now a debug
command for toggling whether or not it should show the reachable area.